requirements.txt
google-cloud-storage
google-cloud-compute